# Improve delete properties performance by replace DOMDocument with xml_parse