--- a/tests/Jackalope/Transport/DoctrineDBAL/ClientTest.php
+++ b/tests/Jackalope/Transport/DoctrineDBAL/ClientTest.php
@@ -12,6 +12,7 @@ use PDO;
 use PHPCR\PropertyType;
 use PHPCR\Query\QOM\QueryObjectModelConstantsInterface;
 use PHPCR\Query\QueryInterface;
+use PHPCR\SimpleCredentials;
 use PHPCR\Util\NodeHelper;
 use PHPCR\Util\PathHelper;
 use PHPCR\Util\QOM\QueryBuilder;
@@ -238,6 +239,34 @@ class ClientTest extends FunctionalTestCase
         $this->session->save();
 
         $this->addToAssertionCount(1);
+    }
+
+    public function testDeleteProperties()
+    {
+        $root = $this->session->getNode('/');
+        $node = $root->addNode('delete-properties');
+        for ($i = 0; $i <= 1000; $i++) {
+            $node->setProperty('property-'.$i, 'value-'.$i);
+        }
+
+        $this->session->save();
+        $this->assertSame(1002, \count($node->getProperties()));
+
+        for ($i = 501; $i <= 1000; $i++) {
+            $node->setProperty('property-'.$i, null);
+        }
+
+        $this->session->save();
+        $this->session->refresh(false);
+        $node = $this->session->getNode('/delete-properties');
+
+        for ($i = 0; $i <= 1000; $i++) {
+            $this->assertSame(
+                $i < 501,
+                $node->hasProperty('property-'.$i),
+                'Unexpected result for property "property-'.$i.'"'
+            );
+        }
     }
 
     public function testPropertyLengthAttribute()