@alexander-schranz 2023-11-24
Thx for the reviews.

@dbu we will next week do some more profiling on some projects and then I will continue here with some fixes. Pushed this already so we can test it on various projects via:

```json
    "repositories": [
        {
            "type": "vcs",
            "url": "git@github.com:alexander-schranz/jackalope-doctrine-dbal.git"
        }
    ],
    "require": {
    
        "jackalope/jackalope-doctrine-dbal": "dev-feature/improve-delete-properties-performance  as 1.10.0",
        
    }
```