@alexander-schranz 2023-12-11
Applied to suggestions the pull request so is ready :) Also did again more testing with special chars and the result is 100% the same for old and new implementations:

<img width="867" alt="Bildschirmfoto 2023-12-11 um 09 53 37" src="https://github.com/jackalope/jackalope-doctrine-dbal/assets/1698337/33dfbc7b-7756-4011-acb8-98678a995c86">